refactor(engine,cli): decompose runner and workspace-run modules past the 700-line bar
- engine/runner.py: EngineFlowRunner mixin owns the step/flow execution
  lifecycle (markers, memory tracking, authoritative final save,
  post-processing, db cleanup, observer/render gate); flow.py keeps the
  data/state/build spine at 321 lines. run_step now reconciles the
  self-archive reader on both success and exception paths.
- cli/command_handlers/workspace_run.py: worker-call construction, suffix
  preflight, and outcome reconciliation move out of project.py (499
  lines). The preflight now catches dependency-check exceptions and covers
  the sizer runtime sentinel (src/sizer_os.tcl), and
  archive_own_step_logs guards descriptor-duplication failures so a broken
  setup releases the guard and restores fds.
Regressions pin direct run_step downgrade, preflight rejection without
mutation, and setup-failure recovery.
