+++ chipcompiler/runtime/log_stream.py
@@ -395,12 +395,14 @@ def archive_own_step_logs(workspace_dir, *, echo: bool = True):
     sys.stdout.flush()
     sys.stderr.flush()
     flush_cstdio()
-    real_stdout = os.dup(1)
-    real_stderr = os.dup(2)
+    real_stdout = -1
+    real_stderr = -1
     read_fd = -1
     write_fd = -1
     stream = None
     try:
+        real_stdout = os.dup(1)
+        real_stderr = os.dup(2)
         read_fd, write_fd = os.pipe()
         os.dup2(write_fd, 1)
         os.dup2(write_fd, 2)
@@ -423,8 +425,12 @@ def archive_own_step_logs(workspace_dir, *, echo: bool = True):
     except BaseException:
         # A setup failure must not poison later runs: restore any redirected
         # descriptors, close what was opened, and release the guard.
-        os.dup2(real_stdout, 1)
-        os.dup2(real_stderr, 2)
+        if real_stdout >= 0:
+            with suppress(OSError):
+                os.dup2(real_stdout, 1)
+        if real_stderr >= 0:
+            with suppress(OSError):
+                os.dup2(real_stderr, 2)
         if write_fd >= 0:
             with suppress(OSError):
                 os.close(write_fd)
@@ -434,8 +440,12 @@ def archive_own_step_logs(workspace_dir, *, echo: bool = True):
         elif read_fd >= 0:
             with suppress(OSError):
                 os.close(read_fd)
-        os.close(real_stdout)
-        os.close(real_stderr)
+        if real_stdout >= 0:
+            with suppress(OSError):
+                os.close(real_stdout)
+        if real_stderr >= 0:
+            with suppress(OSError):
+                os.close(real_stderr)
         _SELF_ARCHIVE_ACTIVE = False
         raise
     try:
